Ajustado conflitos
